Split actions.ts into actions.server.ts and actions.client.ts
