+++ app/mcp/actions.client.ts
@@ -1,0 +1,1 @@
+export * from "./actions.base";
