+++ app/mcp/actions.server.ts
@@ -1,0 +1,2 @@
+"use server";
+export * from "./actions.base";
